Restructure whole repository with each chapter as individual sub-projects.
Instead of having directories for data, results, and sections at the root
of the project, with chapter names as subdirectories, have each chapter
at the root with data, results, and sections subdirectories.  This will
make it easier to later convert a chapter into a separate project (for
a paper).  It will also make it easier to include the histone catalogue
which we are now working as a paper in a separate git repo.

Also, less messy. We don't usually look at everything at the same time,
we only work in a chapter at a time.